requirements.txt
autogluon.timeseries
pandas
numpy
matplotlib